Update compiler flag setup to use modern cross-distro conventions
These new macros were introduced in RPM 4.15 and offer a cross-distro
interface to accessing various language macros and setting the macros
in the build environment.

As we had older analogues, map them to the new cross-distro form and
mark all the old macros for deprecation.

As we are technically defining the macros here (and redefining them
in an RPM 4.15 based environment), this is backwards compatible to
OpenMandriva Lx 4 (using RPM 4.14).